+++ client/src/app.html
@@ -1,0 +1,15 @@
+<div class="content mdl-layout mdl-js-layout mdl-layout--fixed-header">
+	<header class="mdl-layout__header">
+		<div class="mdl-layout__header-row">
+			<span class="mdl-layout-title">Angular 2 REST Website</span>
+			<div class="mdl-layout-spacer"></div>
+			<nav class="mdl-navigation">
+				<a class="mdl-navigation__link" [routerLink]="links.items">Items</a>
+				<a class="mdl-navigation__link" [routerLink]="links.widgets">Widgets</a>
+			</nav>
+		</div>
+	</header>
+	<main class="mdl-layout__content page-content">
+		<router-outlet></router-outlet>
+	</main>
+</div>
