+++ client/src/app.ts
@@ -1,142 +1,20 @@
-import {Component, Input, Output, EventEmitter, ChangeDetectionStrategy} from 'angular2/core';
-import {ItemsService, Item, AppStore} from './items';
-import {Observable} from 'rxjs/Observable';
-import {Store} from '@ngrx/store';
+import {Component, Input, Output, EventEmitter, OnInit} from 'angular2/core';
+import {RouteConfig, ROUTER_DIRECTIVES, Router} from 'angular2/router';
+import {Items} from './items/items.component';
+import {Widgets} from './widgets/widgets.component';
 
-//-------------------------------------------------------------------
-// ITEMS-LIST
-//-------------------------------------------------------------------
-@Component({
-  selector: 'items-list',
-  template: `
-  <div *ngFor="#item of items" (click)="selected.emit(item)"
-    class="item-card mdl-card mdl-shadow--2dp">
-    <div class="mdl-card__title">
-      <h2 class="mdl-card__title-text">{{item.name}}</h2>
-    </div>
-    <div class="mdl-card__supporting-text">
-      {{item.description}}
-    </div>
-    <div class="mdl-card__menu">
-      <button (click)="deleted.emit(item); $event.stopPropagation();"
-        class="mdl-button mdl-button--icon mdl-js-button mdl-js-ripple-effect">
-        <i class="material-icons">close</i>
-      </button>
-    </div>
-  </div>
-  `
-})
-class ItemList {
-  @Input() items: Item[];
-  @Output() selected = new EventEmitter();
-  @Output() deleted = new EventEmitter();
-}
-
-//-------------------------------------------------------------------
-// ITEM DETAIL
-//-------------------------------------------------------------------
-@Component({
-  selector: 'item-detail',
-  template: `
-  <div class="item-card mdl-card mdl-shadow--2dp">
-    <div class="mdl-card__title">
-      <h2 class="mdl-card__title-text" *ngIf="selectedItem.id">Editing {{originalName}}</h2>
-      <h2 class="mdl-card__title-text" *ngIf="!selectedItem.id">Create New Item</h2>
-    </div>
-    <div class="mdl-card__supporting-text">
-      <form novalidate>
-          <div class="mdl-textfield mdl-js-textfield">
-            <label>Item Name</label>
-            <input [(ngModel)]="selectedItem.name"
-              placeholder="Enter a name"
-              class="mdl-textfield__input" type="text">
-          </div>
-
-          <div class="mdl-textfield mdl-js-textfield">
-            <label>Item Description</label>
-            <input [(ngModel)]="selectedItem.description"
-              placeholder="Enter a description"
-              class="mdl-textfield__input" type="text">
-          </div>
-      </form>
-    </div>
-    <div class="mdl-card__actions">
-        <button type="submit" (click)="cancelled.emit(selectedItem)"
-          class="mdl-button mdl-js-button mdl-js-ripple-effect">Cancel</button>
-        <button type="submit" (click)="saved.emit(selectedItem)"
-          class="mdl-button mdl-js-button mdl-button--colored mdl-js-ripple-effect">Save</button>
-    </div>
-  </div>
-  `
-})
-class ItemDetail {
-  @Input('item') _item: Item;
-  originalName: string;
-  selectedItem: Item;
-  @Output() saved = new EventEmitter();
-  @Output() cancelled = new EventEmitter();
-
-  set _item(value: Item){
-    if (value) this.originalName = value.name;
-	  this.selectedItem = Object.assign({}, value);
-  }
-}
-
-//-------------------------------------------------------------------
-// MAIN COMPONENT
-//-------------------------------------------------------------------
 @Component({
   selector: 'my-app',
-  providers: [],
-  template: `
-  <div class="mdl-cell mdl-cell--6-col">
-    <items-list [items]="items | async"
-      (selected)="selectItem($event)" (deleted)="deleteItem($event)">
-    </items-list>
-  </div>
-  <div class="mdl-cell mdl-cell--6-col">
-    <item-detail
-      (saved)="saveItem($event)" (cancelled)="resetItem($event)"
-      [item]="selectedItem | async">Select an Item</item-detail>
-  </div>
-  `,
-  directives: [ItemList, ItemDetail],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  template: require('./app.html'),
+  directives: [ROUTER_DIRECTIVES]
 })
+@RouteConfig([
+  {path: '/items', name: 'Items', component: Items, useAsDefault: true},
+  {path: '/widgets', name: 'Widgets', component: Widgets}
+])
 export class App {
-  items: Observable<Array<Item>>;
-  selectedItem: Observable<Item>;
-
-  constructor(private itemsService: ItemsService, private store: Store<AppStore>) {
-    this.items = itemsService.items;
-    this.selectedItem = store.select('selectedItem');
-    this.selectedItem.subscribe(v => console.log(v));
-
-    itemsService.loadItems();
-  }
-
-  resetItem() {
-    let emptyItem: Item = {id: null, name: '', description: ''};
-    this.store.dispatch({type: 'SELECT_ITEM', payload: emptyItem});
-  }
-
-  selectItem(item: Item) {
-    this.store.dispatch({type: 'SELECT_ITEM', payload: item});
-  }
-
-  saveItem(item: Item) {
-    this.itemsService.saveItem(item);
-
-    // Generally, we would want to wait for the result of `itemsService.saveItem`
-    // before resetting the current item.
-    this.resetItem();
-  }
-
-  deleteItem(item: Item) {
-    this.itemsService.deleteItem(item);
-
-    // Generally, we would want to wait for the result of `itemsService.deleteItem`
-    // before resetting the current item.
-    this.resetItem();
+  links = {
+    items: ['Items'],
+    widgets: ['Widgets']
   }
 }
