+++ client/src/common/models/appstore.model.ts
@@ -1,0 +1,6 @@
+import {Item} from './item.model';
+
+export interface AppStore {
+  items: Item[];
+  selectedItem: Item;
+};
