+++ client/src/common/models/item.model.ts
@@ -1,0 +1,5 @@
+export interface Item {
+  id: number;
+  name: string;
+  description: string;
+};
