+++ client/src/common/models/widget.model.ts
@@ -1,0 +1,5 @@
+export interface Widget {
+  id: number,
+  name: string,
+  price: number
+}
