+++ client/src/common/stores/selectedItem.store.ts
@@ -1,0 +1,8 @@
+export const selectedItem = (state: any = null, {type, payload}) => {
+  switch (type) {
+    case 'SELECT_ITEM':
+      return payload;
+    default:
+      return state;
+  }
+};
