--- client/vendor.ts
@@ -1,6 +1,0 @@
-// Polyfills
-import 'es6-promise';
-import 'es6-shim';
-import 'reflect-metadata';
-import 'zone.js/dist/zone-microtask';
-import 'zone.js/dist/long-stack-trace-zone';
