+++ karma.conf.js
@@ -1,73 +1,19 @@
-// Karma configuration
-// Generated on Mon Jan 25 2016 09:33:05 GMT-0700 (MST)
-
 module.exports = function(config) {
-  var testWebpackConfig = require('./webpack.test.config.js');
+  var webpackConfig = require('./webpack.test.config.js');
+
   config.set({
-
-    // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: '',
-
-
-    // frameworks to use
-    // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['jasmine'],
-
-
-    // list of files / patterns to load in the browser
-    files: [
-      { pattern: 'spec-bundle.js', watched: false }
-    ],
-
-
-    // list of files to exclude
-    exclude: [
-    ],
-
+    frameworks: [ 'jasmine' ],
+    files: [ { pattern: 'spec-bundle.js', watched: false } ],
+    preprocessors: { 'spec-bundle.js': ['webpack', 'sourcemap'] },
+    webpack: webpackConfig,
     webpackServer: { noInfo: true },
-
-    // preprocess matching files before serving them to the browser
-    // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {
-      'spec-bundle.js': ['webpack']
-    },
-
-    webpack: testWebpackConfig,
-
-    // test results reporter to use
-    // possible values: 'dots', 'progress'
-    // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-    reporters: ['spec'],
-
-
-    // web server port
+    reporters: [ 'spec' ],
     port: 9876,
-
-
-    // enable / disable colors in the output (reporters and logs)
     colors: true,
-
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     logLevel: config.LOG_INFO,
-
-
-    // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: true,
-
-
-    // start these browsers
-    // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    browsers: ['PhantomJS'],
-
-
-    // Continuous Integration mode
-    // if true, Karma captures browsers, runs the tests and exits
-    singleRun: false,
-
-    // Concurrency level
-    // how many browser should be started simultaneous
-    concurrency: Infinity
-  })
-}
+    autoWatch: false,
+    browsers: [ 'PhantomJS' ],
+    singleRun: true
+  });
+};
