+++ package.json
@@ -1,100 +1,46 @@
 {
-  "name": "ng2-rest-website",
+  "name": "fem-ng2-ngrx-app",
   "version": "0.0.1",
   "description": "Simple REST website in Angular 2 and NGRX",
   "repository": "https://github.com/onehungrymind/fem-ng2-ngrx-app",
   "scripts": {
-    "clean": "npm cache clean && rimraf node_modules doc typings coverage dist",
-    "clean:dist": "rimraf dist",
-    "preclean:install": "npm run clean",
-    "clean:install": "npm set progress=false && npm install",
-    "preclean:start": "npm run clean",
-    "clean:start": "npm start",
-    "watch": "npm run watch:dev",
-    "watch:dev": "webpack --watch --progress --profile --colors --display-error-details --display-cached",
-    "watch:dev:hmr": "npm run watch:dev -- --hot",
-    "watch:test": "npm run test -- --auto-watch --no-single-run",
-    "watch:prod": "webpack --watch --config webpack.prod.config.js --progress --profile --colors --display-error-details --display-cached",
-    "build": "npm run build:dev",
-    "prebuild:dev": "npm run clean:dist",
-    "build:dev": "webpack --progress --profile --colors --display-error-details --display-cached",
-    "prebuild:prod": "npm run clean:dist",
-    "build:prod": "webpack --config webpack.prod.config.js --progress --profile --colors --display-error-details --display-cached --bail",
     "client": "npm run client:dev:hmr",
-    "client:dev": "webpack-dev-server --inline --progress --profile --colors --watch --display-error-details --display-cached  --port 3001 --content-base client/public",
+    "client:dev": "webpack-dev-server --inline --colors --watch --display-error-details --display-cached  --port 3001 --content-base client/public",
     "client:dev:hmr": "npm run client:dev -- --hot",
-    "client:prod": "http-server dist --cors",
     "server": "json-server --watch server/api/db.json",
-    "webdriver:update": "webdriver-manager update",
-    "webdriver:start": "webdriver-manager start",
-    "lint": "./node_modules/.bin/tslint src/**/*.ts",
-    "pree2e": "npm run webdriver:update -- --standalone",
-    "e2e": "protractor",
-    "e2e:live": "protractor --elementExplorer",
     "test": "karma start",
-    "ci": "npm run e2e && npm run test",
-    "docs": "typedoc --options typedoc.json --exclude **/*.spec.ts ./src/",
     "start": "concurrent \"npm run server\" \"npm run client\"",
     "typings-install": "typings install",
-    "postinstall": "npm run typings-install",
-    "preversion": "npm test",
-    "version": "npm run build",
-    "postversion": "git push && git push --tags"
+    "postinstall": "npm run typings-install"
   },
   "dependencies": {
     "@ngrx/store": "^1.3.3",
-    "angular2": "2.0.0-beta.9",
-    "es6-promise": "^3.1.2",
-    "es6-shim": "^0.33.3",
-    "es7-reflect-metadata": "^1.6.0",
-    "rxjs": "5.0.0-beta.2",
-    "zone.js": "0.5.15"
+    "angular2": "^2.0.0-beta.9",
+    "es6-promise": "^3.0.2",
+    "es6-shim": "^0.33.13",
+    "reflect-metadata": "0.1.2",
+    "rxjs": "^5.0.0-beta.2",
+    "systemjs": "^0.19.8",
+    "zone.js": "^0.5.15"
   },
   "devDependencies": {
     "awesome-typescript-loader": "^0.16.0-rc.0",
-    "compression-webpack-plugin": "^0.3.0",
     "concurrently": "^2.0.0",
-    "copy-webpack-plugin": "^1.1.1",
-    "css-loader": "^0.23.1",
-    "es6-promise-loader": "^1.0.1",
-    "exports-loader": "^0.6.3",
-    "expose-loader": "^0.7.1",
-    "file-loader": "^0.8.5",
-    "html-webpack-plugin": "^2.9.0",
-    "http-server": "^0.9.0",
-    "ie-shim": "^0.1.0",
-    "imports-loader": "^0.6.5",
-    "istanbul-instrumenter-loader": "^0.2.0",
-    "json-loader": "^0.5.4",
-    "karma": "^0.13.21",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-coverage": "^0.5.3",
+    "core-js": "^2.2.0",
+    "jasmine-core": "^2.4.1",
+    "karma": "^0.13.22",
+    "karma-coverage": "^0.5.5",
     "karma-jasmine": "^0.3.7",
-    "karma-mocha-reporter": "^2.0.0",
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "1.7.0",
-    "ng2lint": "0.0.10",
-    "parse5": "^1.3.2",
-    "phantomjs-polyfill": "0.0.2",
-    "phantomjs-prebuilt": "^2.1.4",
-    "protractor": "^3.1.1",
-    "raw-loader": "0.5.1",
-    "remap-istanbul": "^0.5.1",
-    "rimraf": "^2.5.2",
+    "karma-spec-reporter": "0.0.24",
+    "karma-webpack": "^1.7.0",
+    "phantomjs-prebuilt": "^2.1.5",
+    "raw-loader": "^0.5.1",
     "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.0",
-    "ts-helper": "0.0.1",
-    "ts-node": "^0.5.5",
-    "tslint": "^3.5.0",
-    "tslint-loader": "^2.1.3",
-    "typedoc": "^0.3.12",
-    "typescript": "~1.8.7",
-    "typings": "^0.7.8",
-    "url-loader": "^0.5.7",
+    "typescript": "^1.7.3",
     "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-md5-hash": "^0.0.5"
+    "webpack-dev-server": "^1.14.1"
   },
   "author": "Lukas Ruebbelke",
   "license": "MIT"
