+++ webpack.test.config.js
@@ -1,77 +1,23 @@
-/*
- * Helper: root(), and rootDir() are defined at the bottom
- */
-var path = require('path');
-var webpack = require('webpack');
-
-/*
- * Config
- */
-module.exports = {
-  devtool: 'inline-source-map',
-  debug: true,
-
-  entry: {
-    'app': './client/bootstrap.ts' // our angular app
-  },
-
-  resolve: {
-    // ensure loader extensions match
-    extensions: ['','.ts','.js','.json', '.css', '.html']
-  },
-
-  module: {
-    preLoaders: [ { test: /\.ts$/, loader: 'tslint-loader' } ],
-    loaders: [
-      // Support for .ts files.
-      {
-        test: /\.ts$/,
-        loader: 'ts-loader',
-        query: {
-          'ignoreDiagnostics': [
-            2403, // 2403 -> Subsequent variable declarations
-            2300, // 2300 Duplicate identifier
-            2374, // 2374 -> Duplicate number index signature
-            2375  // 2375 -> Duplicate string index signature
-          ]
-        },
-        exclude: [/node_modules/ ]
-      },
-
-      // Support for *.json files.
-      { test: /\.json$/,  loader: 'json-loader' },
-
-      // Support for CSS as raw text
-      { test: /\.css$/,   loader: 'raw-loader' },
-
-      // support for .html as raw text
-      { test: /\.html$/,  loader: 'raw-loader' },
-    ],
-    noParse: [ /zone\.js\/dist\/.+/, /angular2\/bundles\/.+/ ]
-  },
-
-  // Other module loader config
-  tslint: {
-    emitErrors: false,
-    failOnHint: false
-  },
-  // our Webpack Development Server config
-  devServer: {
-    historyApiFallback: true,
-    contentBase: 'client/public',
-    publicPath: '/__build__'
-  }
-
-};
-
-// Helper functions
+var webpack = require('webpack'),
+  path = require('path');
 
 function root(args) {
   args = Array.prototype.slice.call(arguments, 0);
   return path.join.apply(path, [__dirname].concat(args));
 }
 
-function rootNode(args) {
-  args = Array.prototype.slice.call(arguments, 0);
-  return root.apply(path, ['node_modules'].concat(args));
-}
+module.exports = {
+  // the difference from the main webpack config
+  // is that we have removed "debug", "entry", "output", and "devServer" configs
+  // and moved the source map loader from "preloaders" to "loaders"
+  devtool: 'inline-source-map',
+  resolve: {
+    extensions: ['', '.ts', '.js']
+  },
+  module: {
+    loaders: [
+      { test: /\.js$/, loader: "source-map-loader", exclude: [ root('node_modules/rxjs') ]},
+      { test: /\.ts$/, loader: 'awesome-typescript-loader' }
+    ]
+  }
+};
